+++ scripts/main.py
@@ -5,7 +5,7 @@ from gymnasium.wrappers import RecordVideo
 from bomberman_rl import ScoreRewardWrapper, RestrictedKeysWrapper, FlattenWrapper
 
 from argparsing import parse
-from learning_agent.agent import Agent as MyLearningAgent
+from learning_agent.agent import Agent
 
 class DummyAgent:
     def setup(self):
@@ -14,30 +14,30 @@ class DummyAgent:
     def act(self, *args, **kwargs):
         return None
 
-def loop(env, agent, args):
-    state, info = env.reset()
-    terminated, truncated, quit = False, False, False
-
-    while not (terminated or truncated):
-        if args.user_play:
-            action, quit = env.unwrapped.get_user_action()
-            while action is None and not quit:
-                time.sleep(0.1)  # wait for user action or quit
+def loop(env, agent, args, n_episodes=100):
+    for i in range(n_episodes):
+        state, info = env.reset()
+        terminated, truncated, quit = False, False, False
+        while not (terminated or truncated):
+            if args.user_play:
                 action, quit = env.unwrapped.get_user_action()
-        else:
-            action, quit = agent.act(state), env.unwrapped.get_user_quit()
+                while action is None and not quit:
+                    time.sleep(0.1)  # wait for user action or quit
+                    action, quit = env.unwrapped.get_user_action()
+            else:
+                action, quit = agent.act(state), env.unwrapped.get_user_quit()
 
-        if quit:
-            env.close()
-            return None
-        else:
-            new_state, _, terminated, truncated, info = env.step(action)
-            if args.train:
-                agent.game_events_occurred(state, action, new_state, info["events"])
-            state = new_state
+            if quit:
+                env.close()
+                return None
+            else:
+                new_state, _, terminated, truncated, info = env.step(action)
+                if args.train:
+                    agent.game_events_occurred(state, action, new_state, info["events"])
+                state = new_state
 
-    if args.train:
-        agent.end_of_round()
+        if args.train:
+            agent.end_of_round()
 
     if not args.no_gui:
         quit = env.unwrapped.get_user_quit()
@@ -52,8 +52,7 @@ def provideAgent(passive: bool):
     if passive:
         return DummyAgent()
     else:
-        agent = MyLearningAgent()
-        agent.setup()
+        agent = Agent()
         return agent
 
 def main(argv=None):
@@ -62,9 +61,9 @@ def main(argv=None):
 
     # Notice that you can not use wrappers in the tournament!
     # However, you might wanna use this example interface to kickstart your experiments
-    env = RestrictedKeysWrapper(env, keys=["self_pos", "coins", "walls"])
-    env = FlattenWrapper(env)
     env = ScoreRewardWrapper(env)
+    env = RestrictedKeysWrapper(env, keys=["self_pos"])
+    env = FlattenWrapper(env)
     if args.video:
         env = RecordVideo(env, video_folder=args.video, name_prefix=args.match_name)