+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - e.g. add ```--passive``` mode
     - for further changes, see ```--help```
 - negative score for killing oneself
+- interface of environment internal agents to equal environment external agent
 ### Fixed
 - prevent PyGame window from rendering for render modes other than ```"human"```
 - action format: environment only accepts action inputs from proper action space (i.e. numbers)